cli: refactor: move options to own module, split Cli & App types
